TOA able to communicate with management app.

+++ src/app/api/run/[runId]/approve/route.ts
@@ -1,5 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server'
-import { deleteFile } from '@/app/utils'
+import { deleteFile, generateAuthorizationHeaders, UPLOAD_DIR } from '@/app/utils'
+import path from 'path'
+import fs from 'fs'
 
 // TODO This component will do two things
 //  1. Post the file to the management API
@@ -12,8 +14,31 @@ export const POST = async (req: NextRequest, { params }: { params: { runId: stri
     if (!runId) {
         return NextResponse.json({ error: 'Missing runId' }, { status: 400 })
     }
+    const filePath = path.join(UPLOAD_DIR, runId)
+    if (fs.existsSync(filePath)) {
+    
+        const fileContent = await fs.promises.readFile(filePath, 'utf-8')
+        console.error('fileContent', fileContent)
+        /// TODO Create a parsable formData by the endpoint
+        const data  = new FormData()
+        data.append('file', new File([Uint8Array.from(fileContent.split(''), c => c.charCodeAt(0))], runId, { type: 'text/csv' }))
+        console.error('data', data)
+        const endpoint = `${process.env.MANAGEMENT_APP_API_URL}/api/run/${process.env.MANAGEMENT_APP_MEMBER_ID}/results`
+        const request = new Request(endpoint, {
+            method: 'POST',
+            body: data,
+            headers:{ ...generateAuthorizationHeaders(),
+                'Content-Type': 'multipart/form-data'
+            },
+        })
 
-    await deleteFile(runId)
+        const response  = await fetch(request)
+        if (!response.ok) {
+            return NextResponse.json({ error: 'Unable to post file' }, { status: 500 })
+        }
+
+        await deleteFile(runId)
+    }
 
     return NextResponse.json({
         success: true,
